requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.4.0
python-dotenv==1.0.1
pytest==8.3.2
# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
alembic>=1.12.0
# AI 브리핑 (개미 마을 수석 이장) — 사용 시만 설치
openai>=1.0.0
anthropic>=0.18.0
# 실시간 시세·뉴스 (market_data)
yfinance>=0.2.40
feedparser>=6.0.11
# 스케줄 브리핑 (매일 9시·17시)
apscheduler>=3.10.0
